Refine component API
