+++ core/src/lib.rs
@@ -6,10 +6,12 @@
 use bevy::core::FixedTimestep;
 use bevy::prelude::*;
 
+pub use constraints::RotationConstraints;
 pub use ext::*;
 pub use gravity::Gravity;
 pub use velocity::{AxisAngle, Velocity};
 
+mod constraints;
 pub mod ext;
 mod gravity;
 pub mod utils;
@@ -275,43 +277,6 @@ impl Default for PhysicMaterial {
         Self {
             restitution: Self::PERFECTLY_INELASTIC_RESTITUTION,
             density: 1.0,
-        }
-    }
-}
-
-/// Component that restrict the rotations caused by forces
-///
-/// Note that angular velocity may still be applied programmatically.
-#[derive(Debug, Copy, Clone, Reflect)]
-pub struct RotationConstraints {
-    /// Set to true to prevent rotations around the x axis
-    pub allow_x: bool,
-
-    /// Set to true to prevent rotations around the y axis
-    pub allow_y: bool,
-
-    /// Set to true to prevent rotations around the Z axis
-    pub allow_z: bool,
-}
-
-impl Default for RotationConstraints {
-    fn default() -> Self {
-        Self {
-            allow_x: true,
-            allow_y: true,
-            allow_z: true,
-        }
-    }
-}
-
-impl RotationConstraints {
-    /// Lock rotations around all axes
-    #[must_use]
-    pub fn lock() -> Self {
-        Self {
-            allow_x: false,
-            allow_y: false,
-            allow_z: false,
         }
     }
 }
